Speed and code improvements

+++ RCube/ClusterShadingFragmentFactory.cpp
@@ -20,18 +20,18 @@ FragmentFactory::FragmentFactory()
 {
 	masks.resize(48);
 
-	for (int k = 0; k < 16; ++k)
+	for (uint64_t k = 0; k < 16; ++k)
 	{
-		int b = k % 4;
-		int a = k / 4;
+		uint64_t b = k % 4;
+		uint64_t a = k / 4;
 
 		uint64_t one = 1;
 		uint64_t x_segment = 0;
 		uint64_t y_segment = 0;
 		uint64_t z_segment = 0;
 
-		for (int l = a; l <= b; ++l)
-			for (int m = 0; m < 16; ++m)
+		for (uint64_t l = a; l <= b; ++l)
+			for (uint64_t m = 0; m < 16; ++m)
 			{
 				x_segment |= one << ((l / 2 % 2 + m / 8 % 2 * 2) * 16 + (l % 2 + m / 4 % 2 * 2) * 4 + (m % 4));
 				y_segment |= one << ((l / 2 % 2 * 2 + m / 8 % 2) * 16 + (l % 2 * 2 + m / 4 % 2) * 4 + (m % 4));
@@ -46,9 +46,9 @@ FragmentFactory::FragmentFactory()
 
 uint64_t FragmentFactory::coverage(int& x1, int& x2, int& y1, int& y2, int& z1, int& z2) const
 {
-	uint64_t x_segment = masks[0 + x1 * 4 + x2];
-	uint64_t y_segment = masks[16 + y1 * 4 + y2];
-	uint64_t z_segment = masks[32 + z1 * 4 + z2];
+	uint64_t x_segment = masks[0 + (uint64_t)x1 * 4 + (uint64_t)x2];
+	uint64_t y_segment = masks[16 + (uint64_t)y1 * 4 + (uint64_t)y2];
+	uint64_t z_segment = masks[32 + (uint64_t)z1 * 4 + (uint64_t)z2];
 	uint64_t coverage = x_segment & y_segment & z_segment;
 	return coverage;
 }
